+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 __5.0.3 - 2026-08-17__
 * Added code to harden the js save action
+* Added event `JsTrackerCustom.getCustomJsFilePath` so the location of the custom tracker JavaScript file can be changed by other plugins
+* Added a warning when the custom JavaScript is stored outside of the plugin directory while an unused file within the plugin directory is still added to the JavaScript tracker
 
 __5.0.2__
 * Added missing license file